feat(middleware): add new congregation auth check middleware
